requirements.txt
asgiref==3.11.0
Django==6.0.1
django-phonenumber-field==8.4.0
phonenumberslite==9.0.21
psycopg2==2.9.11
sqlparse==0.5.5
